+++ src/DRPLoadingSpinner.m
@@ -14,7 +14,8 @@
 
 @property BOOL isAnimating;
 @property NSUInteger colorIndex;
-@property CAShapeLayer *circleLayer;
+@property CAShapeLayer *foregroundRailLayer;
+@property CAShapeLayer *backgroundRailLayer;
 @property BOOL isFirstCycle;
 
 @end
@@ -47,8 +48,9 @@
 }
 
 - (void)setup {
-    self.circleLayer = [[CAShapeLayer alloc] init];
-
+    self.foregroundRailLayer = [[CAShapeLayer alloc] init];
+    self.backgroundRailLayer = [[CAShapeLayer alloc] init];
+    
     self.drawCycleDuration = 0.75;
     self.rotationCycleDuration = 1.5;
     self.staticArcLength = 0;
@@ -57,20 +59,25 @@
     self.lineWidth = 2.;
     self.opaque = NO;
     self.backgroundColor = [UIColor clearColor];
-
+    
+    self.drawTimingFunction = [DRPLoadingSpinnerTimingFunction easeInOut];
+    
     self.colorSequence = @[
         [UIColor redColor],
         [UIColor orangeColor],
         [UIColor purpleColor],
         [UIColor blueColor]
     ];
+    
+    self.backgroundRailLayer.fillColor = [UIColor clearColor].CGColor;
+    self.backgroundRailLayer.strokeColor = [UIColor clearColor].CGColor;
+    
+    self.foregroundRailLayer.fillColor = [UIColor clearColor].CGColor;
+    self.foregroundRailLayer.anchorPoint = CGPointMake(.5, .5);
+    self.foregroundRailLayer.strokeColor = self.colorSequence.firstObject.CGColor;
+    self.foregroundRailLayer.hidden = YES;
 
-    self.circleLayer.fillColor = [UIColor clearColor].CGColor;
-    self.circleLayer.anchorPoint = CGPointMake(.5, .5);
-    self.circleLayer.strokeColor = self.colorSequence.firstObject.CGColor;
-    self.circleLayer.hidden = YES;
-
-    self.circleLayer.actions = @{
+    self.foregroundRailLayer.actions = @{
         @"lineWidth": [NSNull null],
         @"strokeEnd": [NSNull null],
         @"strokeStart": [NSNull null],
@@ -91,27 +98,36 @@
 - (void)setFrame:(CGRect)frame {
     [super setFrame:frame];
     [self refreshCircleFrame];
-//    [self insertPlane:self.circleLayer];
+//    [self insertPlane:self.foregroundRailLayer];
 }
 
 - (void)setStaticArcLength:(CGFloat)staticArcLength {
     _staticArcLength = staticArcLength;
 
     if (!self.isAnimating) {
-        self.circleLayer.hidden = NO;
-        self.circleLayer.strokeColor = self.colorSequence.firstObject.CGColor;
-        self.circleLayer.strokeStart = 0;
-        self.circleLayer.strokeEnd = [self proportionFromArcLengthRadians:staticArcLength];
-        self.circleLayer.transform = CATransform3DRotate(CATransform3DIdentity, -M_PI_2, 0, 0, 1);
+        self.foregroundRailLayer.hidden = NO;
+        self.foregroundRailLayer.strokeColor = self.colorSequence.firstObject.CGColor;
+        self.foregroundRailLayer.strokeStart = 0;
+        self.foregroundRailLayer.strokeEnd = [self proportionFromArcLengthRadians:staticArcLength];
+        self.foregroundRailLayer.transform = CATransform3DRotate(CATransform3DIdentity, -M_PI_2, 0, 0, 1);
     }
 }
 
 - (void)setLineWidth:(CGFloat)lineWidth {
-    self.circleLayer.lineWidth = lineWidth;
+    self.foregroundRailLayer.lineWidth = lineWidth;
+    self.backgroundRailLayer.lineWidth = lineWidth;
 }
 
 - (CGFloat)lineWidth {
-    return self.circleLayer.lineWidth;
+    return self.foregroundRailLayer.lineWidth;
+}
+
+- (void)setBackgroundRailColor:(UIColor *)backgroundRailColor {
+    self.backgroundRailLayer.strokeColor = backgroundRailColor.CGColor;
+}
+
+- (UIColor *)backgroundRailColor {
+    return [UIColor colorWithCGColor:self.backgroundRailLayer.strokeColor];
 }
 
 #pragma mark - Layout
@@ -143,38 +159,45 @@
     CGFloat sideLen = MIN(self.layer.frame.size.width, self.layer.frame.size.height) - (2 * self.lineWidth);
     CGFloat xOffset = ceilf((self.frame.size.width - sideLen) / 2.0);
     CGFloat yOffset = ceilf((self.frame.size.height - sideLen) / 2.0);
-
-    self.circleLayer.frame = CGRectMake(xOffset, yOffset, sideLen, sideLen);
-    self.circleLayer.path = [UIBezierPath bezierPathWithOvalInRect:CGRectMake(0, 0, sideLen, sideLen)].CGPath;
+    
+    self.foregroundRailLayer.frame = CGRectMake(xOffset, yOffset, sideLen, sideLen);
+    self.foregroundRailLayer.path = [UIBezierPath bezierPathWithOvalInRect:CGRectMake(0, 0, sideLen, sideLen)].CGPath;
+    
+    self.backgroundRailLayer.frame = CGRectMake(xOffset, yOffset, sideLen, sideLen);
+    self.backgroundRailLayer.path = [UIBezierPath bezierPathWithOvalInRect:CGRectMake(0, 0, sideLen, sideLen)].CGPath;
 }
 
 - (void)layoutSublayersOfLayer:(CALayer *)layer {
     [super layoutSublayersOfLayer:layer];
-
-    if (!self.circleLayer.superlayer) {
-        [self.layer addSublayer:self.circleLayer];
+    
+    if (!self.backgroundRailLayer.superlayer) {
+        [self.layer addSublayer:self.backgroundRailLayer];
+    }
+    
+    if (!self.foregroundRailLayer.superlayer) {
+        [self.layer addSublayer:self.foregroundRailLayer];
     }
 }
 
 #pragma mark - Animation control
 - (void)startAnimating {
     [self stopAnimating];
-    self.circleLayer.hidden = NO;
-
+    self.foregroundRailLayer.hidden = NO;
+    
     self.isAnimating = YES;
     self.isFirstCycle = YES;
-    self.circleLayer.strokeEnd = [self proportionFromArcLengthRadians:self.minimumArcLength];
-
+    self.foregroundRailLayer.strokeEnd = [self proportionFromArcLengthRadians:self.minimumArcLength];
+    
     self.colorIndex = 0;
-    self.circleLayer.strokeColor = [self.colorSequence[self.colorIndex] CGColor];
-
-    [self addAnimationsToLayer:self.circleLayer reverse:NO rotationOffset:-M_PI_2];
+    self.foregroundRailLayer.strokeColor = [self.colorSequence[self.colorIndex] CGColor];
+    
+    [self addAnimationsToLayer:self.foregroundRailLayer reverse:NO rotationOffset:-M_PI_2];
 }
 
 - (void)stopAnimating {
     self.isAnimating = NO;
-    self.circleLayer.hidden = YES;
-    [self.circleLayer removeAllAnimations];
+    self.foregroundRailLayer.hidden = YES;
+    [self.foregroundRailLayer removeAllAnimations];
 }
 
 #pragma mark - Auto Layout
@@ -224,6 +247,7 @@
         CABasicAnimation *rotationAnimation = [CABasicAnimation animationWithKeyPath:@"transform.rotation.z"];
         rotationAnimation.fromValue = @(rotationStartRadians);
         rotationAnimation.toValue = @(rotationStartRadians + (2 * M_PI));
+        rotationAnimation.timingFunction = [DRPLoadingSpinnerTimingFunction linear];
         rotationAnimation.duration = self.rotationCycleDuration;
         rotationAnimation.repeatCount = CGFLOAT_MAX;
         rotationAnimation.fillMode = kCAFillModeForwards;
@@ -244,7 +268,7 @@
 
     strokeAnimation.delegate = self;
     strokeAnimation.fillMode = kCAFillModeForwards;
-    strokeAnimation.timingFunction = [CAMediaTimingFunction functionWithName:kCAMediaTimingFunctionEaseInEaseOut];
+    strokeAnimation.timingFunction = self.drawTimingFunction;
     [CATransaction setAnimationDuration:strokeDuration];
 
     [layer removeAnimationForKey:@"stroke"];
@@ -255,7 +279,7 @@
 
 - (void)advanceColorSequence {
     self.colorIndex = (self.colorIndex + 1) % self.colorSequence.count;
-    self.circleLayer.strokeColor = [self.colorSequence[self.colorIndex] CGColor];
+    self.foregroundRailLayer.strokeColor = [self.colorSequence[self.colorIndex] CGColor];
 }
 
 - (CGFloat)proportionFromArcLengthRadians:(CGFloat)radians {
@@ -268,9 +292,9 @@
         BOOL isStrokeStart = [basicAnim.keyPath isEqualToString:@"strokeStart"];
         BOOL isStrokeEnd = [basicAnim.keyPath isEqualToString:@"strokeEnd"];
 
-        CGFloat rotationOffset = fmodf([[self.circleLayer.presentationLayer valueForKeyPath:@"transform.rotation.z"] floatValue], 2 * M_PI);
-        [self addAnimationsToLayer:self.circleLayer reverse:isStrokeEnd rotationOffset:rotationOffset];
-
+        CGFloat rotationOffset = fmodf([[self.foregroundRailLayer.presentationLayer valueForKeyPath:@"transform.rotation.z"] floatValue], 2 * M_PI);
+        [self addAnimationsToLayer:self.foregroundRailLayer reverse:isStrokeEnd rotationOffset:rotationOffset];
+        
         if (isStrokeStart) {
             [self advanceColorSequence];
         }