+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 1.1.0
+* Added the ability to set the timing function of the draw cycle. This allows you to have more control over the subtlety or loudness of the animation.
+* Added a background rail and the ability to set its color. This defaults to clear, so your existing animations won't be affected.
+
 ### 1.0.2
 * Silenced a noisy log message